refactor(ui): change the ui-bottom-sheet to a controlled component
