+++ packages/ui/src/components/ui-bottom-sheet.tsx
@@ -1,21 +1,19 @@
-import type { ReactNode } from 'react'
+import type { ComponentProps, ReactNode } from 'react'
 
-import { Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger } from './sheet.tsx'
+import { Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle } from './sheet.tsx'
 
 export function UiBottomSheet({
   children,
   description,
   title,
-  trigger,
-}: {
+  ...props
+}: ComponentProps<typeof Sheet> & {
   children: ReactNode
   description: ReactNode
   title: ReactNode
-  trigger: ReactNode
 }) {
   return (
-    <Sheet>
-      <SheetTrigger asChild>{trigger}</SheetTrigger>
+    <Sheet {...props}>
       <SheetContent className="sm:-translate-x-1/2 w-full sm:left-1/2 sm:w-[400px]" side="bottom">
         <SheetHeader>
           <SheetTitle>{title}</SheetTitle>
